+++ changelog.md
@@ -4,6 +4,7 @@
 
 * Firmware version 2.0 support.
 * Clock synchronization between OpenBCI board and this driver
+* `.info` property which contains verified board information parsed from the OpenBCI's `softReset()` response.
 
 ### Breaking Changes